+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [v1.5.0] - 2024-07-22
+
 ### Added
 
 - The `whoami` command displays the current Git branch.
@@ -80,7 +82,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## v1.0.0 - 2017-04-03
 - First release.
 
-[Unreleased]: https://github.com/raphaelstolt/git-user-bend/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/raphaelstolt/git-user-bend/compare/v1.5.0...HEAD
+[v1.5.0]: https://github.com/raphaelstolt/git-user-bend/compare/v1.4.0...v1.5.0
 [v1.4.0]: https://github.com/raphaelstolt/git-user-bend/compare/v1.3.1...v1.4.0
 [v1.3.1]: https://github.com/raphaelstolt/git-user-bend/compare/v1.3.0...v1.3.1
 [v1.3.0]: https://github.com/raphaelstolt/git-user-bend/compare/v1.2.2...v1.3.0